updates to different classes, before character update, maybe ints?

+++ app/src/main/java/com/example/wallproject/Model/Dungeon/Characters/Attack.kt
@@ -1,6 +1,6 @@
 package com.example.wallproject.Model.Dungeon.Characters
 
-data class Attack(var attackDmg : Double, var abilityPower : Double, var luck : Double){
+data class Attack(var attackDmg : Int, var abilityPower : Int, var luck : Int){
