+++ app/src/main/java/com/example/wallproject/Model/Dungeon/Characters/Defense.kt
@@ -1,6 +1,6 @@
 package com.example.wallproject.Model.Dungeon.Characters
 
-data class Defense(var armor : Double, var magicResist : Double){
+data class Defense(var armor : Int, var magicResist : Int){
